requirements.txt
beautifulsoup4
python-dotenv
requests
lxml
pymongo
pydantic
fastapi
uvicorn